# feat(wallet): use optimised balance scanner + expanded tokens list