--- a/components/brave_wallet/browser/blockchain_list_parser.cc
+++ b/components/brave_wallet/browser/blockchain_list_parser.cc
@@ -16,6 +16,7 @@
 #include "base/strings/stringprintf.h"
 #include "brave/components/brave_wallet/browser/blockchain_list_schemas.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
+#include "brave/components/brave_wallet/common/brave_wallet.mojom.h"
 #include "brave/components/brave_wallet/common/common_utils.h"
 #include "brave/components/brave_wallet/common/solana_utils.h"
 #include "brave/components/brave_wallet/common/value_conversion_utils.h"
@@ -33,6 +34,11 @@ bool ParseResultFromDict(const base::Value::Dict* response_dict,
   }
   *output_val = *val;
   return true;
+}
+
+bool ParseOptionalBoolFromDict(const base::Value::Dict* response_dict,
+                               const std::string& key) {
+  return response_dict->FindBool(key).value_or(false);
 }
 
 std::optional<double> ParseNullableStringAsDouble(const base::Value& value) {
@@ -205,35 +211,36 @@ void AddTokenToMaps(const blockchain_lists::Token& token,
 
 }  // namespace
 
-bool ParseTokenList(const std::string& json,
-                    TokenListMap* token_list_map,
-                    mojom::CoinType coin) {
+bool ParseTokenList(const std::string& json, TokenListMap* token_list_map) {
   DCHECK(token_list_map);
 
   // {
-  //  "0x0D8775F648430679A709E98d2b0Cb6250d2887EF": {
-  //    "name": "Basic Attention Token",
-  //    "logo": "bat.svg",
-  //    "erc20": true,
-  //    "symbol": "BAT",
-  //    "decimals": 18
-  //  },
-  //  "0x06012c8cf97BEaD5deAe237070F9587f8E7A266d": {
-  //    "name": "Crypto Kitties",
-  //    "logo": "CryptoKitties-Kitty-13733.svg",
-  //    "erc20": false,
-  //    "erc721": true,
-  //    "symbol": "CK",
-  //    "decimals": 0
-  //  },
-  //  "0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984": {
-  //    "name": "Uniswap",
-  //    "logo": "uni.svg",
-  //    "erc20": true,
-  //    "symbol": "UNI",
-  //    "decimals": 18,
-  //    "chainId": "0x1"
-  //  }
+  //   "0x1": {
+  //     "0xdac17f958d2ee523a2206206994597c13d831ec7": {
+  //       "name": "Tether",
+  //       "symbol": "usdt",
+  //       "coingeckoId": "tether",
+  //       "decimals": 6,
+  //       "logo":
+  //       "https://coin-images.coingecko.com/coins/images/325/large/Tether.png"
+  //     },
+  //     "0xb8c77482e45f1f44de1745f52c74426c631bdd52": {
+  //       "name": "BNB",
+  //       "symbol": "bnb",
+  //       "coingeckoId": "binancecoin",
+  //       "decimals": 18,
+  //       "logo":
+  //       "https://coin-images.coingecko.com/coins/images/825/large/bnb-icon2_2x.png"
+  //     },
+  //     "0x06012c8cf97BEaD5deAe237070F9587f8E7A266d": {
+  //       "name": "Crypto Kitties",
+  //       "logo": "CryptoKitties-Kitty-13733.svg",
+  //       "erc20": false,
+  //       "erc721": true,
+  //       "symbol": "CK",
+  //       "decimals": 0
+  //     }
+  //   }
   // }
 
   std::optional<base::Value> records_v =
@@ -244,70 +251,69 @@ bool ParseTokenList(const std::string& json,
     return false;
   }
 
-  const auto& response_dict = records_v->GetDict();
-  for (const auto blockchain_token_value_pair : response_dict) {
-    auto blockchain_token = mojom::BlockchainToken::New();
-    blockchain_token->contract_address = blockchain_token_value_pair.first;
-    const auto* blockchain_token_value =
-        blockchain_token_value_pair.second.GetIfDict();
-    if (!blockchain_token_value) {
-      return false;
-    }
-
-    blockchain_token->is_erc20 =
-        blockchain_token_value->FindBool("erc20").value_or(false);
-
-    blockchain_token->is_erc721 =
-        blockchain_token_value->FindBool("erc721").value_or(false);
-
-    blockchain_token->is_nft = blockchain_token->is_erc721;
-
-    if (!ParseResultFromDict(blockchain_token_value, "symbol",
-                             &blockchain_token->symbol)) {
+  const auto& chain_dict = records_v->GetDict();
+  // Iterate through chain IDs
+  for (const auto chain_pair : chain_dict) {
+    const std::string& chain_id = chain_pair.first;
+    const auto* tokens_dict = chain_pair.second.GetIfDict();
+    if (!tokens_dict) {
       continue;
     }
-    if (!ParseResultFromDict(blockchain_token_value, "name",
-                             &blockchain_token->name)) {
-      return false;
-    }
-    ParseResultFromDict(blockchain_token_value, "logo",
-                        &blockchain_token->logo);
 
-    std::optional<int> decimals_opt =
-        blockchain_token_value->FindInt("decimals");
-    if (decimals_opt) {
+    // Determine coin type based on chain_id
+    mojom::CoinType coin = chain_id == mojom::kSolanaMainnet
+                               ? mojom::CoinType::SOL
+                               : mojom::CoinType::ETH;
+
+    // Iterate through tokens in this chain
+    for (const auto token_pair : *tokens_dict) {
+      auto blockchain_token = mojom::BlockchainToken::New();
+      blockchain_token->contract_address = token_pair.first;
+      const auto* token_info = token_pair.second.GetIfDict();
+      if (!token_info) {
+        continue;
+      }
+
+      // Parse required fields
+      if (!ParseResultFromDict(token_info, "symbol",
+                               &blockchain_token->symbol) ||
+          !ParseResultFromDict(token_info, "name", &blockchain_token->name)) {
+        continue;
+      }
+
+      // Parse optional fields
+      ParseResultFromDict(token_info, "logo", &blockchain_token->logo);
+      ParseResultFromDict(token_info, "coingeckoId",
+                          &blockchain_token->coingecko_id);
+      blockchain_token->is_erc721 =
+          ParseOptionalBoolFromDict(token_info, "erc721");
+
+      // Determining is_nft from blockchain list is not supported for Solana
+      blockchain_token->is_nft = blockchain_token->is_erc721;
+      bool is_token2022 = ParseOptionalBoolFromDict(token_info, "token2022");
+
+      std::optional<int> decimals_opt = token_info->FindInt("decimals");
+      if (!decimals_opt) {
+        continue;
+      }
       blockchain_token->decimals = *decimals_opt;
-    } else {
-      continue;
-    }
 
-    // chain_id is only optional for ETH mainnet token lists.
-    blockchain_token->chain_id = "0x1";
-    if (!ParseResultFromDict(blockchain_token_value, "chainId",
-                             &blockchain_token->chain_id) &&
-        coin != mojom::CoinType::ETH) {
-      continue;
-    }
-
-    ParseResultFromDict(blockchain_token_value, "coingeckoId",
-                        &blockchain_token->coingecko_id);
-
-    blockchain_token->coin = coin;
-    blockchain_token->visible = true;
-
-    if (IsSPLToken(blockchain_token)) {
-      bool is_token2022 =
-          blockchain_token_value->FindBool("token2022").value_or(false);
+      // Set default values
+      blockchain_token->chain_id = chain_id;
+      blockchain_token->coin = coin;
+      blockchain_token->visible = true;
+      blockchain_token->is_erc20 =
+          coin == mojom::CoinType::ETH && !blockchain_token->is_nft;
+      blockchain_token->is_erc1155 = false;
       blockchain_token->spl_token_program =
-          is_token2022 ? mojom::SPLTokenProgram::kToken2022
-                       : mojom::SPLTokenProgram::kToken;
-    } else {
-      blockchain_token->spl_token_program =
-          mojom::SPLTokenProgram::kUnsupported;
-    }
+          IsSPLToken(blockchain_token)
+              ? (is_token2022 ? mojom::SPLTokenProgram::kToken2022
+                              : mojom::SPLTokenProgram::kToken)
+              : mojom::SPLTokenProgram::kUnsupported;
 
-    (*token_list_map)[GetTokenListKey(coin, blockchain_token->chain_id)]
-        .push_back(std::move(blockchain_token));
+      (*token_list_map)[GetTokenListKey(coin, chain_id)].push_back(
+          std::move(blockchain_token));
+    }
   }
 
   return true;