… components/brave_wallet/browser/blockchain_list_parser_unittest.cc
… components/brave_wallet/browser/brave_wallet_constants.cc
… components/brave_wallet/browser/brave_wallet_constants.h
… components/brave_wallet/browser/json_rpc_service.cc
… components/brave_wallet/browser/json_rpc_service.h
… components/brave_wallet/browser/json_rpc_service_unittest.cc
… components/brave_wallet_ui/common/constants/magics.ts
… components/brave_wallet_ui/common/slices/endpoints/token_balances.endpoints.ts