## Token list expansion

We're now using a tokens list derived from coingecko for all preloaded chains including Solana. The list is updated here on a nightly basis: https://github.com/brave/token-lists/blob/main/data/v1/coingecko-top5000.json

* Use remote images (hosted on coingecko) for token logos instead of wallet data files.
   @brave/crypto-wallets-ios Please check if this is fine.
* List now expanded to top 5000 coins on Coingecko. Note that (for example) Tether counts as one coin on all supported chains, so actual number of tokens in this new list is more than 5000.
* No noticeable performance regression due to the new tokens list size.

## Performance Improvements

Courtesy https://github.com/brave/evm-scanner

### Page (re)load CPU usage

Slight improvement in CPU usage observed.

|Before|After|
|-|-|
| <img src="https://github.com/user-attachments/assets/3b4b39b7-d2e8-4613-823c-549d656e3942" /> |<img src="https://github.com/user-attachments/assets/fbe85bfe-00c0-448e-b3ad-356558b59403" /> |

### Mojo calls

Significant reduction in number of Mojo/network calls observed.

| Mojo method                             | Before | After | Reduction |
|-------------------------------|-------:|------:|----------:|
| `jsonRpcService.getBalance`  | 72    | 30    |   58%     |
| `jsonRpcService.getSPLTokenAccountBalance`     | 160    | 0     |     100%   |
| `jsonRpcService.getSPLTokenBalances`     | 0    | 4     |   -100%  |
| `jsonRpcService.getERC20TokenBalance`          | 18     | 12     |   33%    |
| `jsonRpcService.getERC20TokenBalances`        | 30    | 42 |  -100% |
| `jsonRpcService.getSolanaBalance` | 4 | 4 | 0% |
| `jsonRpcService.getNftBalances`                | 66     | 10     |   85%     |
| **Total** | **350** | **102** | **71%** |

Resolves https://github.com/brave/brave-browser/issues/44018
Resolves https://github.com/brave/brave-browser/issues/44019

<!-- CI-related labels that can be applied to this PR:
* CI/run-audit-deps (1) - check for known npm/cargo vulnerabilities (audit_deps)
* CI/run-network-audit (1) - run network-audit
* CI/run-upstream-tests - run Chromium unit and browser tests on Linux and Windows (otherwise only on Linux)
* CI/run-linux-arm64, CI/run-macos-x64, CI/run-windows-arm64, CI/run-windows-x86 - run builds that would otherwise be skipped
* CI/skip - do not run CI builds (except noplatform)
* CI/skip-linux-x64, CI/skip-android, CI/skip-macos-arm64, CI/skip-ios, CI/skip-windows-x64 - skip CI builds for specific platforms
* CI/skip-upstream-tests - do not run Chromium unit, or browser tests (otherwise only on Linux)
* CI/skip-all-linters - do not run presubmit and lint checks
* CI/storybook-url (1) - deploy storybook and provide a unique URL for each build

(1) applied automatically when some files are changed (see: https://github.com/brave/brave-core/blob/master/.github/labeler.yml)
-->

## Submitter Checklist:

- [x] I confirm that no [security/privacy review is needed](https://github.com/brave/brave-browser/wiki/Security-reviews) and no other type of reviews are needed, or that I have [requested](https://github.com/brave/reviews/issues/new/choose) them
- [x] There is a [ticket](https://github.com/brave/brave-browser/issues) for my issue
- [x] Used Github [auto-closing keywords](https://docs.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue) in the PR description above
- [x] Wrote a good [PR/commit description](https://google.github.io/eng-practices/review/developer/cl-descriptions.html)
- [x] Squashed any review feedback or "fixup" commits before merge, so that history is a record of what happened in the repo, not your PR
- [x] Added appropriate labels (`QA/Yes` or `QA/No`; `release-notes/include` or `release-notes/exclude`; `OS/...`) to the associated issue
- [x] Checked the PR locally:
  * `npm run test -- brave_browser_tests`, `npm run test -- brave_unit_tests` [wiki](https://github.com/brave/brave-browser/wiki/Tests)
  * `npm run presubmit` [wiki](https://github.com/brave/brave-browser/wiki/Presubmit-checks), `npm run gn_check`, `npm run tslint`
- [x] Ran `git rebase master` (if needed)

## Reviewer Checklist:

- [ ] A security review [is not needed](https://github.com/brave/brave-browser/wiki/Security-reviews), or a link to one is included in the PR description
- [ ] New files have MPL-2.0 license header
- [ ] Adequate test coverage exists to prevent regressions
- [ ] Major classes, functions and non-trivial code blocks are well-commented
- [ ] Changes in component dependencies are properly reflected in `gn`
- [ ] Code follows the [style guide](https://chromium.googlesource.com/chromium/src/+/HEAD/styleguide/c++/c++.md)
- [ ] Test plan is specified in PR before merging

## After-merge Checklist:

- [ ] The associated issue milestone is set to the smallest version that the
  changes has landed on
- [ ] All relevant documentation has been updated, for instance:
  - [ ] https://github.com/brave/brave-browser/wiki/Deviations-from-Chromium-(features-we-disable-or-remove)
  - [ ] https://github.com/brave/brave-browser/wiki/Proxy-redirected-URLs
  - [ ] https://github.com/brave/brave-browser/wiki/Fingerprinting-Protections
  - [ ] https://github.com/brave/brave-browser/wiki/Brave%E2%80%99s-Use-of-Referral-Codes
  - [ ] https://github.com/brave/brave-browser/wiki/Web-Compatibility-Exceptions-in-Brave
  - [ ] https://github.com/brave/brave-browser/wiki/QA-Guide
  - [ ] https://github.com/brave/brave-browser/wiki/P3A

## Test Plan:

Ensure token balances are correctly fetched, including NFTs, across all networks.