@brave-builds 2025-02-07
A Storybook has been deployed to [preview UI for the latest push](https://build-artifacts.brave.com/brave-browser-build-pr/f/wallet/cg-tokens-list/storybook.html)