@StephenHeaps 2025-02-11
> Use remote images (hosted on coingecko) for token logos instead of wallet data files.
> @brave/crypto-wallets-ios Please check if this is fine.

This is fine on iOS. We will [fetch from URL](https://github.com/brave/brave-core/blob/bb97a678d0b5d5c59cf40198388970966552e456/ios/brave-ios/Sources/BraveWallet/Crypto/AssetIconView.swift#L23) if there is [no image stored/available](https://github.com/brave/brave-core/blob/bb97a678d0b5d5c59cf40198388970966552e456/ios/brave-ios/Sources/BraveWallet/Extensions/BraveWalletExtensions.swift#L482-L501).