@brave-builds 2025-02-17
Released in [v1.77.44](https://github.com/brave/brave-browser/releases/tag/v1.77.44)